Rename and move ui element icons for palette to new file
