+++ inst/ui-editor-react/src/ElementsPalette/index.tsx
@@ -2,14 +2,12 @@ import * as React from "react";
 
 import type {
   ShinyUiNames,
-  ShinyUiNode,
   ShinyUiNodeInfo,
 } from "Shiny-Ui-Elements/uiNodeTypes";
 import { shinyUiNodeInfo } from "Shiny-Ui-Elements/uiNodeTypes";
 
-import { useMakeDraggable } from "../DragAndDropHelpers/useMakeDraggable";
-
 import classes from "./styles.module.css";
+import { UiElementIcon } from "./UiElementIcon";
 
 const categoryOrder: string[] = [
   "Inputs",
@@ -45,29 +43,8 @@ export default function ElementsPalette({
   return (
     <div className={classes.elementsPalette}>
       {ui_node_names.map((uiName) => (
-        <ElementOption key={uiName} uiName={uiName} />
+        <UiElementIcon key={uiName} uiName={uiName} />
       ))}
-    </div>
-  );
-}
-
-function ElementOption({ uiName }: { uiName: ShinyUiNames }) {
-  const { iconSrc, title, defaultSettings } = shinyUiNodeInfo[uiName];
-  const node = {
-    uiName,
-    uiArguments: defaultSettings,
-  } as ShinyUiNode;
-
-  const elRef = React.useRef<HTMLDivElement>(null);
-  useMakeDraggable({ ref: elRef, nodeInfo: { node } });
-
-  if (iconSrc === undefined) {
-    return null;
-  }
-  return (
-    <div ref={elRef} className={classes.OptionItem} data-ui-name={uiName}>
-      <img src={iconSrc} alt={title} className={classes.OptionIcon} />
-      <label className={classes.OptionLabel}>{title}</label>
     </div>
   );
 }
